refactor(tasks): make vertex-tasks the single home for the MaybeSend markers
